chore: lint intergrations

+++ pkg/models/mysql.go
@@ -134,7 +134,7 @@ type MySQLResultSet struct {
 }
 
 type PacketHeader struct {
-	PacketLength     uint8 `json:"packet_length,omitempty" yaml:"packet_length,omitempty,flow" bson:"packet_length,omitempty"`
+	PacketLength     uint32 `json:"packet_length,omitempty" yaml:"packet_length,omitempty,flow" bson:"packet_length,omitempty"`
 	PacketSequenceID uint8 `json:"packet_sequence_id,omitempty" yaml:"packet_sequence_id,omitempty,flow" bson:"packet_sequence_id,omitempty"`
 }
